Move SourceMap::Map -> SourceMap
