+++ lib/sprockets/coffee_script_processor.rb
@@ -23,7 +23,7 @@ module Sprockets
         Autoload::CoffeeScript.compile(data, sourceMap: true, sourceFiles: [input[:source_path]])
       end
 
-      map = input[:metadata][:map] | Sprockets::SourceMap::Map.from_json(result['v3SourceMap'])
+      map = input[:metadata][:map] | Sprockets::SourceMap.from_json(result['v3SourceMap'])
       { data: result['js'], map: map }
     end
   end
