+++ lib/sprockets/digest_utils.rb
@@ -80,8 +80,8 @@ module Sprockets
         elsif klass == Encoding
           digest << 'Encoding'
           digest << obj.name
-        elsif klass == SourceMap::Map
-          digest << 'SourceMap::Map'
+        elsif klass == SourceMap
+          digest << 'SourceMap'
           queue << obj.as_json
         else
           raise TypeError, "couldn't digest #{klass}"
