+++ test/test_uglifier_compressor.rb
@@ -10,7 +10,7 @@ class TestUglifierCompressor < MiniTest::Test
       data: "function foo() {\n  return true;\n}",
       cache: Sprockets::Cache.new,
       metadata: {
-        map: Sprockets::SourceMap::Map.new
+        map: Sprockets::SourceMap.new
       }
     }
     output = "function foo(){return!0}"
